pyproject.toml
[project]
name = "seminar2024-pixkit-can"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pygame>=2.6.1",
    "python-can>=4.4.2",
]
